+++ cedar-policy/CHANGELOG.md
@@ -43,6 +43,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - For the `partial-eval` experimental feature: make the return values of
   `RequestBuilder`'s `principal`, `action`, `resource`, `context` and
   `schema` functions `#[must_use]`. (#502)
+- For the `partial-eval` experimental feature: make `RequestBuilder::schema`
+return a `RequestBuilder<&Schema>` so the `RequestBuilder<&Schema>::build`
+method checks the request against the schema provided and the 
+`RequestBuilder<UnsetSchema>::build` method becomes infallible. (#559)
 
 ### Fixed